+++ map/example.cub
@@ -8,15 +8,15 @@ C 225,30,222
 
         1111111111111111111111111
         1000000000110000000000001
-        1011000001110000011100001
-        100100000000000001 10000111111111111111111111
-111111111011000001110000011100001
+   1    1011000001110000011100001
+        100100000000000001 100001
+1111111110110000011100000111000N1
 100000000011000001110111111111111
 11110111111111011100000010001
 11110111111111011101010010001
 11000000110101011100000000001
 10000000000000001100000010001
 10000000000000001101010010001
-11000001110101011111011110N0111
+1100000111010101111101111000111
 11110111 1110101 101111010001
 11111111 1111111 111111111111
