+++ src/utils_mlx.c
@@ -1,0 +1,39 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils_mlx.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: alefranc <alefranc@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/08/30 14:41:08 by alefranc          #+#    #+#             */
+/*   Updated: 2022/08/30 14:41:33 by alefranc         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+void	my_mlx_pixel_put(t_data *data, int x, int y, int color)
+{
+	char	*dst;
+
+	dst = data->addr + (y * data->line_length + x * (data->bits_per_pixel / 8));
+	*(unsigned int *)dst = color;
+}
+
+void	my_mlx_square(t_data *data, int posx, int posy, int color)
+{
+	int	x;
+	int	y;
+
+	y = 0;
+	while (y < TILE)
+	{
+		x = 0;
+		while (x < TILE)
+		{
+			my_mlx_pixel_put(data, posx + x, posy + y, color);
+			x++;
+		}
+		y++;
+	}
+}
